requirements.txt
asgiref==3.3.4
boto3==1.17.54
botocore==1.20.54
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
django-tagging==0.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
importlib-metadata==3.10.1
jmespath==0.10.0
Markdown==3.3.4
packaging==20.9
Pillow==8.2.0
protobuf==3.15.8
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
s3transfer==0.4.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.4.1
